fix: Make like project button clickable in workbench
