+++ apps/web/src/features/project/hub/components/ProjectCard.tsx
@@ -1,9 +1,8 @@
 import {
   useDuplicateProject,
-  useLikeProject,
   useModifyProject,
-  useUnlikeProject,
 } from "@/queries/mutations/projectMutations";
+import { ProjectLikeButton } from "@/features/project/components/ProjectLikeButton.tsx";
 import { ludoNavigation } from "@/constants/ludoNavigation.tsx";
 import { CustomIcon } from "@ludocode/design-system/primitives/custom-icon.tsx";
 import { Avatar } from "@ludocode/design-system/primitives/avatar.tsx";
@@ -18,7 +17,7 @@ import type { ProjectCardResponse } from "@ludocode/types";
 import { qo } from "@/queries/definitions/queries.ts";
 import { useQuery } from "@tanstack/react-query";
 import { ProjectVisibilityMenu } from "./ProjectVisibilityMenu";
-import { ArrowRightIcon, Copy, Heart } from "lucide-react";
+import { ArrowRightIcon, Copy } from "lucide-react";
 import { cn } from "@ludocode/design-system/cn-utils";
 import { Languages } from "@ludocode/types/Project/ProjectFileSnapshot.ts";
 
@@ -323,61 +322,3 @@ function ProjectCopyButton({ projectId, userId }: ProjectRemixButtonProps) {
   );
 }
 
-type ProjectLikeButtonProps = {
-  projectId: string;
-  canLike: boolean;
-  className?: string;
-};
-
-function ProjectLikeButton({
-  projectId,
-  canLike,
-  className,
-}: ProjectLikeButtonProps) {
-  const { data: likeState } = useQuery(qo.projectLike(projectId));
-  const likeProjectMutation = useLikeProject(projectId);
-  const unlikeProjectMutation = useUnlikeProject(projectId);
-
-  const count = likeState?.count ?? 0;
-  const isLikedByMe = likeState?.likedByMe;
-  const isPending =
-    likeProjectMutation.isPending || unlikeProjectMutation.isPending;
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-    event.preventDefault();
-    event.stopPropagation();
-
-    if (!canLike || isPending) return;
-
-    if (isLikedByMe) {
-      unlikeProjectMutation.mutate();
-      return;
-    }
-
-    likeProjectMutation.mutate();
-  };
-
-  return (
-    <button
-      type="button"
-      disabled={!canLike || isPending}
-      onClick={handleClick}
-      className={cn(
-        "flex items-center gap-1 text-ludo-white",
-        !canLike || isPending
-          ? "hover:cursor-not-allowed"
-          : "hover:cursor-pointer",
-        className,
-      )}
-    >
-      <Heart
-        fill={isLikedByMe ? "#F87171" : "none"}
-        className={cn(
-          "h-4",
-          isLikedByMe ? "text-[#F87171]" : "text-ludo-white",
-        )}
-      />
-      <span className="text-sm leading-none tabular-nums">{count}</span>
-    </button>
-  );
-}
